espFlaskApi/requirements.txt
Flask==3.1.0
numpy==2.2.2
sounddevice==0.5.1
websocket_client==1.8.0
python-dotenv==1.0.1